Eater Bug fix and xp/hr
Finer grain mouse movement in eater to stop the overWatch from
destroying it.

Added xp/hr.

Made a local state in the overwatch such that if the state changes
during the mouse movement check it does not release the mouse (if
owned), just to limit the logs - had no effect I can see.

Added Settings state in overwatch in case of mouse being used to change
settings.